+++ packages/fighting-design/tag/index.ts
@@ -4,12 +4,14 @@ import { install } from '../_utils'
 
 export const FTag = install(Tag)
 
-export type FTagInstance = InstanceType<typeof Tag>
+export type TagInstance = InstanceType<typeof Tag>
+
+export type { TagPropsType } from './src/tag'
+
+export type { TagSizeType, TagType } from './src/interface'
 
 declare module 'vue' {
   export interface GlobalComponents {
     FTag: typeof FTag
   }
 }
-
-export type { tagSizeType, tagTypeType } from './src/interface'