+++ packages/fighting-design/tag/src/interface.d.ts
@@ -1,6 +1,6 @@
-export type tagSizeType = 'large' | 'middle' | 'small' | 'mini'
+export type TagSizeType = 'large' | 'middle' | 'small' | 'mini'
 
-export type tagTypeType = 'default' | 'primary' | 'success' | 'danger' | 'warning'
+export type TagType = 'default' | 'primary' | 'success' | 'danger' | 'warning'
 
 export interface handleCloseInterface {
   (evt: Event): void