+++ packages/fighting-design/toolbar-item/index.ts
@@ -4,7 +4,9 @@ import { install } from '../_utils'
 
 export const FToolbarItem = install(ToolbarItem)
 
-export type FToolbarItemInstance = InstanceType<typeof ToolbarItem>
+export type ToolbarItemInstance = InstanceType<typeof ToolbarItem>
+
+export type { ToolbarItemPropsType } from './src/toolbar-item'
 
 declare module 'vue' {
   export interface GlobalComponents {