+++ packages/fighting-design/toolbar-item/src/interface.d.ts
@@ -1,3 +1,0 @@
-export interface handleClickInterface {
-  (event: MouseEvent): void
-}