+++ packages/fighting-design/toolbar/index.ts
@@ -4,12 +4,14 @@ import { install } from '../_utils'
 
 export const FToolbar = install(Toolbar)
 
-export type FToolbarInstance = InstanceType<typeof Toolbar>
+export type ToolbarInstance = InstanceType<typeof Toolbar>
+
+export type { ToolbarPropsType } from './src/toolbar'
+
+export type { ToolbarType, ToolbarClickEmitInterface } from './src/interface'
 
 declare module 'vue' {
   export interface GlobalComponents {
     FToolbar: typeof FToolbar
   }
 }
-
-export type { toolbarTypeType, toolbarFixedStyleInterface } from './src/interface'