+++ packages/fighting-design/toolbar/src/interface.d.ts
@@ -1,18 +1,6 @@
-export type toolbarTypeType = 'large' | 'middle' | 'small' | 'mini'
+export type ToolbarType = 'large' | 'middle' | 'small' | 'mini'
 
-export interface handleClickInterface {
-  // (evt: MouseEvent & { path: HTMLElement[] }): void
-  (evt: MouseEvent): void
-}
-
-export interface clickEmitInterface {
+export interface ToolbarClickEmitInterface {
   evt: MouseEvent
   key: string | undefined
-}
-
-export interface toolbarFixedStyleInterface {
-  top?: string
-  button?: string
-  right?: string
-  left?: string
 }